Moved all components into component folder
