+++ src/components/NavbarBottom.js
@@ -1,22 +1,22 @@
 import React, { Component } from 'react';
-import './navbarBottom.css';
+import '../styles/navbarBottom.css';
 
 class NavbarBottom extends Component {
     render() {
         return (
-            
-            <ul className = "nav-two">
-                <span className = "leftbar">
+
+            <ul className="nav-two">
+                <span className="leftbar">
                     <li><button>Board Name</button></li>
                     <li><button className="star"><i className="far fa-star"></i></button></li>
                     <li><button>Team</button></li>
                     <li><button>Privacy</button></li>
                 </span>
-                <span className = "rightbar">
+                <span className="rightbar">
                     <li><button>... Show Menu</button></li>
                 </span>
             </ul>
-            
+
         );
     }
 }
